+++ FeedModule/FeedModule/FeedCache/FeedCachePolicy.swift
@@ -1,0 +1,25 @@
+//
+//  FeedCachePolicy.swift
+//  FeedModule
+//
+//  Created by Omran Khoja on 2/10/22.
+//
+
+import Foundation
+
+internal final class FeedCachePolicy {
+    private init() {}
+    
+    private static let calendar = Calendar(identifier: .gregorian)
+    
+    private static var maxCacheAgeInDays: Int {
+        return 7
+    }
+    
+    internal static func validate(_ timestamp: Date, from currentDate: Date) -> Bool {
+        guard let maxCacheAge = calendar.date(byAdding: .day, value: maxCacheAgeInDays, to: timestamp) else {
+            return false
+        }
+        return currentDate < maxCacheAge
+    }
+}
